+++ test/translator/mock_ai_service.dart
@@ -1,0 +1,5 @@
+import 'package:mockito/annotations.dart';
+import 'package:tafsiri/core/services/ai_service.dart';
+
+@GenerateMocks([AiService])
+void main() {}